Add files via upload
Some initial EDA on the wildfires dataset, which contains fires in LA County since 2014. The highlights of this file include determining the distribution of fire frequency, finding the most correlated variables, determining whether fire response and containment are related, and adding a fire duration in days column to the dataset.